#116: dedup newTempDir into appleTest TempDirs helper
Per review-reuse: NSTemporaryDirectory + UUID + per-test cleanup
list was copy-pasted into five appleTest files. Extracted into
TempDirs (slug-parameterised). DeviceKeyPairTest, FileServerTest,
FileServerConcurrencyTest, FileServerPairTest, and
AppleUploadStorageBackendTest now delegate to it.
